Refactoring TimeSeries

+++ aquifolium/include/BTC.h
@@ -181,7 +181,7 @@ template<class T> T diff(CTimeSeries<T> &BTC_p, CTimeSeries<T> &BTC_d);
 template<class T> T diff_abs(CTimeSeries<T> &BTC_p, CTimeSeries<T> &BTC_d);
 template<class T> T diff_log(CTimeSeries<T> &BTC_p, CTimeSeries<T> &BTC_d, double lowlim);
 template<class T> T diff_norm(CTimeSeries<T> &BTC_p, CTimeSeries<T> &BTC_d);
-template<class T> T diff(CTimeSeries<T> BTC_p, CTimeSeries<T> BTC_d, int scale);
+template<class T> T diff(CTimeSeries<T> BTC_p, CTimeSeries<T> BTC_d, double scale);
 template<class T> T diff(CTimeSeries<T> BTC_p, CTimeSeries<T> BTC_d, CTimeSeries<T> Q);
 template<class T> T diff2(CTimeSeries<T> *BTC_p, CTimeSeries<T> BTC_d);
 template<class T> T diff2(const CTimeSeries<T> &BTC_p, CTimeSeries<T> *BTC_d);
